refactor: unclone cycles
Another trivial refactoring

There is a `for _ in something.clone()` patter within the code and some
such clones are called millions of times. Remove it and get an 6.6%
performance gain

```console
time:   [42.805 s 42.876 s 42.931 s]
change: [-6.8468% -6.6651% -6.4971%] (p = 0.00 < 0.05)
Performance has improved.
```

I also removed clones of `BitUint` since it is not necessary to clone
values for conversion to bits.